Prepopulate items to tokenInput.
Gotcha: according to documentation, item should be formed as

```javascript
{
  id: 3,
  name: "John"
}
```

but name is not taken as name, and that item is displayed as `undefined`.

Inspecting internals of item object showed that it contains `id`,
containing an `id` and `value`, `hint` and `queue`, containing name. I
picked `value` and it worked for me. @aratak, you may want to look at it
more careful.